requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
